refactor audit logging and improve failure tracking

+++ src/main/java/com/guts/Guts_IAM/token/service/RefreshTokenService.java
@@ -1,7 +1,9 @@
     package com.guts.Guts_IAM.token.service;
 
+    import com.guts.Guts_IAM.auditlog.action.AuditStatus;
     import com.guts.Guts_IAM.auditlog.model.AuditLog;
     import com.guts.Guts_IAM.auditlog.repository.AuditRepository;
+    import com.guts.Guts_IAM.auditlog.service.AuditLogService;
     import com.guts.Guts_IAM.common.exception.types.TokenNotFoundException;
     import com.guts.Guts_IAM.security.util.hashutil.HashUtil;
     import com.guts.Guts_IAM.security.jwt.dto.JwtResponse;
@@ -10,11 +12,14 @@
     import com.guts.Guts_IAM.token.audit.TokenAuditRepository;
     import com.guts.Guts_IAM.token.refreshtoken.model.RefreshToken;
     import com.guts.Guts_IAM.token.refreshtoken.repository.RefreshTokenRepository;
+    import com.guts.Guts_IAM.user.model.User;
+    import com.guts.Guts_IAM.user.repository.UserRepository;
     import jakarta.servlet.http.HttpServletRequest;
     import lombok.RequiredArgsConstructor;
     import org.springframework.http.HttpStatus;
     import org.springframework.stereotype.Service;
     import org.springframework.web.server.ResponseStatusException;
+    import com.guts.Guts_IAM.auditlog.action.Action;
 
     import java.time.Instant;
     import java.util.Date;
@@ -28,15 +33,40 @@
 
         private final RefreshTokenRepository refreshTokenRepository;
         private final JwtUtils jwtUtils;
-        private final AuditRepository auditRepository;
         private final TokenAuditRepository tokenAuditRepository;
+        private final AuditLogService auditLogService;
+
+
 
         public JwtResponse refreshAccessToken(String refreshTokenStr, HttpServletRequest httpServletRequest) {
-            RefreshToken refreshToken = refreshTokenRepository.findByToken(refreshTokenStr)
-                    .orElseThrow(() -> new TokenNotFoundException("Refresh token not found","NOT_FOUND",HttpStatus.NOT_FOUND));
+            Optional<RefreshToken> refreshToken1 = refreshTokenRepository.findByToken(refreshTokenStr);
+
+            if(refreshToken1.isEmpty()){
+                auditLogService.log(
+                        null,
+                        Action.REFRESH_ACCESS_TOKEN,
+                        "AUTH",
+                        null,
+                        AuditStatus.FAILED,
+                        "Refresh token not found",
+                        httpServletRequest);
+
+                throw new TokenNotFoundException("Refresh token not found", "NOT_FOUND",HttpStatus.NOT_FOUND);
+            }
+
+            RefreshToken refreshToken=refreshToken1.get();
+
+            User user=refreshToken.getUser();
 
             if (refreshToken.getExpiryDate().before(Date.from(Instant.now()))) {
                 refreshTokenRepository.delete(refreshToken);
+                auditLogService.log(user,
+                        Action.REFRESH_ACCESS_TOKEN,
+                        "AUTH",
+                        user.getUserId().toString(),
+                        AuditStatus.EXPIRED,
+                        "Refresh token expired and was revoked",
+                        httpServletRequest);
                 throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Invalid refresh token");
             }
 
@@ -44,26 +74,21 @@
 
             TokenAudit tokenAudit=new TokenAudit();
             tokenAudit.setTokenOwner(refreshToken.getUser().getUserMail());
-            tokenAudit.setAction("REFRESH_TOKEN");
+            tokenAudit.setAction("REFRESH_ACCESS_TOKEN");
             String hashedNewToken= HashUtil.sha256(newAccessToken);
             tokenAudit.setAccessToken(hashedNewToken);
             tokenAuditRepository.save(tokenAudit);
 
-            AuditLog auditLog = new AuditLog();
-            auditLog.setLogAction("REFRESH_TOKEN");
-            auditLog.setUserMail(refreshToken.getUser().getUserMail());
-            auditLog.setResource("AUTH");
-            auditLog.setResourceId(refreshToken.getUser().getUserId().toString());
-            auditLog.setUserId(refreshToken.getUser().getUserId());
-            auditLog.setRoleName(
-                    Optional.ofNullable(refreshToken.getUser().getRoles())
-                            .orElse(Set.of())
-                            .toString()
-            );
-            auditLog.setIpAddress(httpServletRequest.getRemoteAddr());
-            auditLog.setUserAgent(httpServletRequest.getHeader("User-Agent"));
 
-            auditRepository.save(auditLog);
+            auditLogService.log(user,
+                    Action.REFRESH_ACCESS_TOKEN,
+                    "AUTH",
+                    user.getUserId().toString(),
+                    AuditStatus.SUCCESS,
+                    "Access token refreshed successfully",
+                    httpServletRequest);
+
+
             return new JwtResponse(newAccessToken, refreshToken.getToken(), "Bearer");
         }
     }
